requirements.txt
django
djangorestframework
Pillow
django-countries